requirements.txt
Flask==0.12.1
requests==2.21.0
pytest==3.10.1
Flask_Cors==3.0.10
flask_mail==0.9.1
PyJWT==2.0.1
